Increase attempts_before_hard_state for the procfile worker memory
This is a new check and event handler, so wait a while before
considering memory usage a failure.

In the case of Sidekiq specifically, it's probably good to wait a
while after Sidekiq has used a load of memory to increase the chance
that the job that caused the increased memory usage has finished.