Upgrade to Brakeman 4.6 or above
Currently we're running Brakeman 3.5.1 because that's the version that
was installed at the time. Since then there is a newer version of
Brakeman (4.6.1). By changing the string to `~> 4.6` we can ensure that
we have the latest 4.x version but we'll need to manually intervene if
we want version 5 (when it comes out).