Sort the tooltips on the Sidekiq Grafana dashboard
Quite a few of the panels on the dashboard contain a large number of
series, where only a few will have values at a given time. The jobs
processed per second panel is a good example of this, as there's a
series for each job on each machine, but only a few of those will be
active at a time.

Therefore, sort the items in the tooltip to group together the actual
data. Ideally we could just not show empty series in the tooltip, but
I haven't worked out how to do that.