Remove search_benchmark job
This job runs the healthcheck script in search-performance-explorer.
That healthcheck uses a manually-generated set of test data[1] from
Google Analytics which needs to be periodically updated to remain of
value.  But nobody has updated this data since July 2018.

We're moving towards monitoring of query performance using
Elasticsearch's ranking evaluation API, which is not the approach this
benchmark takes, so introducing a process to regularly update the test
data isn't useful.  In which case, there's no point in keeping this
job.

[1] https://github.com/alphagov/search-performance-explorer/blob/master/data/click_model.json