remove custom plek entries for publishing api in AWS staging
As part of the migration of publishing_api to AWS staging, we need to remove the custom PLEK entries which direct AWS nodes to use the publishing api external address since these nodes should use the internal address of publishing_api in AWS.

It should be noted that the Carrenza nodes should still use the external address of the publishing api.

We simplify the logic of configuring the custom env variable by setting it in the govuk/manifest/deploy/config so that these env variables are set on all machines rather than a select few.