Fix gdal_version hieradata for CI agents
This is now a bit disjointed, as Mapit is running in AWS, and so the
AWS hieradata applies, but we still want a matching version of gdal on
the ci_agents in Carrenza.

As the reference no longer works, just hardcode the version. Hopefully
if this changes prior to migrating away from Carrenza, all the
instances of the version will be updated.