Change postgres user to aws_db_admin on all Postgres restores.
When restoring the Publishing API database, govuk_env_sync substitutes
`aws_db_admin` for `postgres` on statements which grant permissions.
This is necessary because of differences in the was Amazon RDS handles
user account vs. ordinary Postgres.

This transformation needs to apply to all database restores into RDS,
not just Publishing API.