Run email-alert-api Postgres backup from primary instead of standby.
Unfortunately the `govuk_env_sync` backup does not run reliably from the
standby. While it is possible in principle to work around this, the
extra complexity, effort and risk involved is unlikely to be worth it
seeing as we anticipate moving from self-hosted Postgres to RDS within
a matter of months.