Fix the trend panel on the Sentry errors dashboard
Previously, it was displaying the average number of errors per hour,
but the title of the panel was "Trend (average over 1d)", which was
completely wrong :(

Now it's summing the errors in an interval, and it's now possible to
change the interval. The title of the panel has also been updated.