Delete the router::assets_origin spec
This is less relevant, and not particularly workable now that
infrastructure is split across AWS and Carrenza, and managed DNS is
used more for routing traffic compared to entries in the /etc/hosts
files on machines.

A part of it was already commented out, so continue and remove the
entire file. The real_data.yaml file is removed as this looks to be
the only use of it.