Add scheduled job for search-api rank evaluation
This will run a search rake task every 3 hours. The rake
task populates graphite with the ranking evaluation scores
for a bunch of queries that we have relevancy judgements for.

These scores are reported in the search relevancy dashboard,
allowing the search team to monitor search relevancy over time.

This also adds a contact group for the search team, so alerts
will go to our Slack channel.

The job runs at 45 mins after the hour because the data sync
which deletes elasticsearch indices runs at 24 mins past the hour
for a few mins, and it'd break the rank eval job if they ran at
exactly the same time. This is quite brittle, but should prevent
this happening until one of them happen at a different time.

Trello: https://trello.com/c/9qZqdAXs