Reload router instead of restart
The router app is a Go application which supports zero downtime restarts
using github.com/alext/tablecloth

Its deploy script knows about this and reloads the app instead of
stopping/starting or restarting it

However govuk::app has a bunch of definitions which notify
govuk::app::service which by default restarts the service.

Allow govuk::app::service to take a restart which overrides the restart
parameter in the Service

Then we can override the restart command for the router app to run
'service router reload' instead, so that it is restarted in the zero
downtime fashion as intended by the application authors.

Also add some specs for govuk::app::service to ensure no regressions

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>