Partially revert "Remove frontend app_hostnames list"
I think routing requests through the backend-lb machines on the
Carrenza side of Production is broken because the entries for
whitehall-frontend are missing from the /etc/hosts file. Restoring the
whitehall-frontend entries here should make the /etc/hosts file
correct.

This seems to be breaking requests for new assets for Whitehall.

This partially reverts commit
005c0a6c84b25073c55d1ca7d6e789344fd0026e.