Add Icinga checks for technical and internal failures
This adds Icinga checks for high numbers of delivery attempts with
a technical or internal failure.

Previously we would check for this using a healthcheck in
email-alert-api but this has a number of problems:

- The healthchecks are no longer machine specific, meaning they can't be
  used correctly by the load balancer to decide where to send traffic.
- The healthchecks have ended up doing a lot of work (including database
  queries) which slow themselves down but also put unnecessary load on
  the database.
- In Icinga, a single healthcheck problem in email-alert-api ends up
  being duplicated for each machine which can be confusing.

These checks work slightly differently to the existing ones in
email-alert-api because they work on an absolute number of failures
rather than a proportion. This is why I've chosen to have very low
thresholds which we can change in the future if we think these are too
low.

The number of technical and internal failures while the app is running
normally is already very low:

https://graphite.publishing.service.gov.uk/render/?width=1053&height=487&_salt=1572948737.847&from=-1hours&target=sum(stats.govuk.app.email-alert-api.*.delivery_attempt.status.technical_failure)&target=sum(stats.govuk.app.email-alert-api.*.delivery_attempt.status.internal_failure)