Stop restarting rsyslog every 5 minutes on most monitoring machines
The bouncer CDN log is only present on a minority of the monitoring
machines, and where it's not present, rsyslog is currently restarted
every 5 minutes. Why, I don't know, but it's not going to magically
make the file appear.

Make this slightly less awful by removing this cron job from the
monitoring machines where bouncer monitoring isn't enabled (as I guess
that means the log file shouldn't be present).