Fix the trend line on the Platform Health Whitehall errors dashboard
Previously each of the data points represented the 30 day average, but
at the mid point of a potentially incomplete 30 day window. This is
far to open to interpretation.

Rather than that, use the movingAverage function to show the 30 day
moving average, which results in more datapoints, but something that I
think is easier to interpret.