Use the combined cert bundle for RDS/DocDB.
The new cert bundle is a superset of the old one and includes the cert
chain for both RDS and DocumentDB.

Also make the cert a Puppet file rather than a template; there's no need
for it to be a template.

There is a reference to rds_cacert.pem in alphagov/router-api. `router-api`
isn't actually using these certs yet because it's still on self-hosted Mongo, but
I'm intentionally not removing the old file yet until that config is
updated, just in case `mongoid` tries to open the file anyway and fails.