Add an alert when we go above our Notify daily limit
I think the graphite query is correct but it's hard to know for sure
until this is deployed to production because we sent very few emails in
the other environments.

I also think we don't need to have the `summarize` method because the
Icinga check will perform a sum over the `from` value (which is set to
24 hours) however without the `summarize` the linked to graph could be
confusing as the individual numbers across the day will be far below
the daily thresholds.