Change the way we resolve an hostname from an instance id
The get_hostname_by_id function get the hostname of an instance
by looking up the value of the key 'PrivateDnsName' for the first
Interface of the instance.
For instances that have more than one interface this can return
unexpected results, it is also unnecessary as the instance has
itself a 'PrivateDnsName' key that will return the result we
expect.