Limit CKAN to 1000 concurrent Postgres connections.
CKAN has recently suffered three incidents where a reindexing job failed
in such as way as to accumulate thousands of concurrent connections to
the shared RDS Postgres, preventing other apps (as well as itself) from
connecting.

This limits the blast radius of such failures to CKAN indexing.

The CKAN / data.gov.uk front end (which runs in GOV.UK PaaS) is not
affected by this change.