Fix Licensify URL used by Draft Frontend.
`draft-frontend` sets `PLEK_HOSTNAME_PREFIX` to prepend `draft-` to the
URLs of its dependencies, so that it uses the special "draft" versions
of those dependencies. Some of these dependencies, including Licensify,
don't have special draft versions. The URLs for those services are
overridden back to their original values by setting
`PLEK_SERVICE_..._URI`.

This adds back the missing override for Licensify, so that
`draft-frontend` doesn't use the `draft-` prefix when connecting to
Licensify.