Don't try to compile Licensify node classes in Carrenza.
There is some logic in `modules/hosts/manifests/purge.pp` which causes
the "does it compile?" tests to fail for these node classes in Carrenza.

Since Licensify no longer runs in Carrenza/UKCloud, avoid running these
tests.

We could add an exception to `purge.pp`, but that would likely just
cause confusion since that module deals with checking `/etc/hosts`
config which is no longer relevant for Licensify now that it's in AWS.