Add infrequent Icinga alert for ad-hoc dashboards
This adds a new alert to check if there are any ad-hoc dashboards in
the Grafana instance associated with the environment. Such dashboards
aren't in version control, tend to be temporary, and accumulate over
time. Having this alert will motivate us to clean them up - currently
our dashboard landscape is quite confusing, because there are so many.