Fix metrics dashboard for Email Alert API
This fixes the new dashboard for Email Alert API to show data from the
new workers that also generate emails as part of processing a content
change. Since there is no data/metrics for messages on production, it's
not currently possible to configure graphs for them, without errors.
Since messages are only a small proportion of the total traffic, it
seems reasonable to leave them out of the graphs for now.

Since 'workload' and 'content changes' are now effectively the same
thing, this replaces the previous 'workload duration' graph with one
for delivery requests and renames the row accordingly.

Finally, this also adds and updates the gauges to reflect the minutely
and daily rate limits in case we hit them.