Set Postgres search path back to default before running sanitise script.
It turns out that pg_dump/pg_restore sets search_path to '' (for
security reasons), meaning that any subsequent SQL statements in the
restore session need to use fully-qualified names when referring to
tables and so on.

Set the search path back to the default so that transformation scripts
don't need to prefix every table name with 'public.'.

Refactor filtered_postgresql_restore yet again in order to accomodate
this while hopefully remaining readable.