Install docker on standalone jenkins instances
This is useful for the search-analytics job, which has recently
started failing because the version of pip we have on the jenkins
instances is so old, it can't install some of its dependencies.

An alternative would be updating pip.  This probably has updating
Python as a pre-requisite, so both would need to be packaged for
Ubuntu 14.04.  We might then discover incompatibilities, where there
are other things running which have broken with a newer version of
Python.

Containers are a nice solution to this: jobs can specify the versions
of dependencies they need, different jobs can use different versions,
and we don't need to manually package stuff.