only try to scrape node_exporter on running hosts
There are some stopped gatling boxes in production, which prometheus
is trying to scrape.  This means that they appear as "down" targets in
prometheus.

I think Prometheus maybe shouldn't try to scrape instances that are
known to be stopped?

Those gatling boxes are a bit weird in any case; they don't have the
same tags as the other boxes in that environment.