Set up proxying for application assets
This lays the foundations for allowing the proxying of asset paths to
application via the www hostname. This sets up the ability to specify a
hash of paths that will be directly proxied to an application before a
request reaches router.

This is done as part of work to enable HTTP performance improvements for
GOV.UK as address in RFC 115 [1]

[1]: https://github.com/alphagov/govuk-rfcs/blob/45e09c088bfde6c1d4dfb68f9bc1c635f63c9a39/rfc-115-enabling-http2-on-govuk.md