Add metrics to show a snapshot of worker activity
Previously we could only know about live Sidekiq activity by connecting
to the Web UI and manually scanning the list of active jobs. This adds
two new graph panels to visualise two new metrics that represent, for
each queue, a snapshot of the number and runtime of its workers.

Having visibility of the live state of Sidekiq will help to diagnose
end-to-end issues that are due to a job getting 'stuck', or running for
longer than expected. Previously we were somewhat blind to this kind of
problem, due to the effort and error-proneness of the Web UI approach.

The new metrics are only available sporadically due to the collection
mechanism used in sidekiq-stats, as is the case for the data shown in
the existing 'queue' panels. Even so, I've tried to make the partial
nature of the data clear by writing 'snapshot' in the graph titles.

Reference: https://github.com/phstc/sidekiq-statsd/blob/master/lib/sidekiq/statsd/server_middleware.rb#L42