Properly remove backend_lb::publishing_api_backend_servers
Some of the configuration was previously removed, but the parameter
needs removing as well, as it's mandatory.

The publishing_api_backend_servers parameter isn't relevant any
longer, as the Publishing API has been migrated to AWS, which doesn't
use the backend_lb node, as it uses AWS services for load balancing.