Add subscription contents Icinga checks
This adds Icinga checks that monitor the number of `email-alert-api`
subscription contents that have a `critical` and `warning` latency
(created more than `x` minutes ago but their associated emails have
not been sent out). They are not machine specific and the data used
lives in stats/gauges/govuk/email-alert-api in Graphite.

We will be alerted if either of these values are bigger than 0.