Fix scheduled publishing dashboard
This dashboard has been broken since the Content Store was moved to AWS
since it refers to a `content-store-1` machine which no longer exists.
I've changed it to aggregate across all the machines and pick the
maximum value. In reality, the values across all the machines should
match as this is not a machine-specific value.