Fix subscription contents Icinga checks
For valid Icinga check configuration, both `warning` and
`critical` arguments are required, but we only need one or
the other for these particular checks (see https://github.com/alphagov/govuk-puppet/pull/10017)

To only trigger a `warning` status for a check (and "disable"
the `critical` status), setting it to a high value that won't be hit
allows us to bypass this issue.

To only trigger a `critical` status for a check, both arguments
can be set to be the same value as `critical` as it takes
precedence.

Although these two checks act as one, the descriptions can’t
be identical as only one check is generated otherwise.

Also includes a small fix to one of the targets which was originally missed.