Add a dashboard for Whitehall Frontend requests
This is related to a goal Platform Health have this year to reduce the
average number of requests being served by Whitehall Frontend.

I decided to put it in its own dashboard because there didn't seem to be
an existing dashboard which was approriate as this dashboard is very
specific to a Platform Health goal. It's also similar to the existing
goal-based dashboard we have:
https://grafana.publishing.service.gov.uk/dashboard/file/platform_health_Q3_19-20_whitehall_errors.json