Remove check_rw_rootfs icinga alert
History lesson: I created this check back in 2012.  Why?

In my role immediately prior to working on GOV.UK, I was working in an
environment where we had some sort of flakiness with storage, which
meant that every now and then a kernel would get confused and remount
the root filesystem as read-only, as a safety measure.  This caused
all sorts of hard-to-debug problems, so a nagios check was put in to
show if this happened again so that people could fix the underlying
problem quickly.

When we first built out nagios for GOV.UK in 2012, we were racking our
brains to try to think of a sensible list of things that we might want
to check as a baseline to make sure our infrastructure was working,
and the idea of this check was recent in my mind.  So I added it.

But GOV.UK has never been deployed to infrastructure that had the kind
of issues that led to read-only root filesystems.  I don't think this
alert has ever fired.

:put_litter_in_its_place: