Add Icinga checks for email-alert-api queue latency
This adds Icinga checks that monitor the latency of the various
Sidekiq queues we have in `email-alert-api`. The tresholds have
been taken from the healthcheck https://github.com/alphagov/email-alert-api/blob/master/app/models/healthcheck/queue_latency.rb The healthcheck will be removed once these checks are in place.

We also want to add other checks on the queues (such as retry size
and queue size), this has been written in a way that should reduce
a lot of duplication in the creation of multiple checks for each
queue.