Add back some prematurely-removed Whitehall hosts entries.
These are the entries for the machines rather than the apps. We don't
want to remove these on migration day because it breaks Puppet on those
Carrenza Whitehall machines and makes it difficult to connect to them
from the jumpbox (because these hosts entries are how the names get
resolved on the jumpbox).

These names should instead be removed after the migration, as a cleanup
task.