Create a Jenkins job to query Athena CDN logs
This is a scheduled task that is due to run in Jenkins 3 times a day to
verify that Athena has fresh data. This check is being added to monitor
that we are a) still receiving data, b) successfully indexing data and
c) said data is queryable. It is being added as a result of a situation
that occurred before Christmas where we discovered that Athena had not
been queryable for 6 weeks due to a JSON error.

This job intends to monitor the govuk_www and govuk_assets databases in
integration, staging and production. It also will monitor the bouncer
database in production only, as we do not have the same bouncer set-up
running in integration and staging.

In the integration and staging environments we have a regular flow of
traffic due to the gor traffic replay that runs from production. This
allows us to be confident that there will be fresh data in these
environments. The gor replay does however get paused during the nightly
data sync so there are periods from 10pm to 4am where there may be 0
traffic from these environments. The job timings are therefore set-up
to be respectful of this and to run 3 times a day at times close to
business hours. I also chose to do the exact same thing
in all 3 environments rather than configuring production different as I
value parity between environments and like to minimise the risk of
problems caused by configuration changes. These checks do run on the
predicate that it would be unheard of for the GOV.UK CDN to go 4 hours
without a request.

I chose to try make a single jenkins job serve for all 3 checks. I don't
think this is a particularly common approach for a GOV.UK Jenkins check
but it feels less messy that 2 or 3 near identical jobs (as is the case
with email alerts). This has meant a few niche techniques such as an
environment variable in the icinga check, linking to a Jenkins job with
a filter parameter and using the parameterized scheduler plugin.

In order to set-up Icinga passive checks for each database I've had to
create an additional defined resource so that we have means to iterate
through the array of databases [1]. Our linter also forced me to do
this in a new file. Hopefully in the not-too-distant future we can
replace this need for an additional defined resource with a loop in the
govuk_jenkins::jobs::athena_fastly_logs_check class.

[1]: https://stackoverflow.com/a/13008766