Remove conditional owner for /data/uploads/publisher
Following migration of whitehall and asset-manager to AWS we still
 have the /data/uploads/publisher dir left on the NFS share.
However the assets user do not exist anymore on the backend
machines. We remove this bit of code setting the owner as a
temporary measure to fix puppet on the backend machines.
Long term solution will be to move the files in
/data/uploads/publisher somewhere else.