Don't tell the Release app about node app deploys
When TAG isn't a TAG, but a branch the Release app doesn't really know
how to make sense of this.

Therefore, just avoid telling the Release app about this. This job is
used to fudge new machines in to the state of existing ones, so it
shouldn't be necessary to tell the Release app anything, as this job
isn't about changing what's released to a given environment.