Rename Whitehall database restore job for consistency.
The govuk_env_sync restore job for the Whitehall MySQL database was the
only one with _mysql in its name. None of the others have the name of
the DBMS in the job name. Since this is something which we may need to
use under pressure in a disaster recovery scenario, let's make it
consistent with the other job names.