Make check_cdn_log_s3_freshness parse times from log filenames.
The initial minimal approach of just seeing whether there are any files
within the closest hour's prefix isn't sufficient. We need to determine
a date/time for the latest log that we can see and then compare that
against the threshold that we're given.

Also change the `-c` argument to specify the threshold log age in
minutes rather than hours, because we get that precision for free now
that we're parsing the times from the filenames instead of relying on
the hour-bucketed prefixes.