Include Python version 3.5 in govuk_python module
This will install `Python 3.5.2` on the CI agent machines and others
that include the `govuk_python` modules (such as `mapit` and `ckan`).

This is currently blocking some of our branches from building
successfully as they require `Python 3.5 or later`, e.g Mapit.

`Python3` seems to be locked on `3.4.3` and some machines are still
using this so we will continue to also install `python3-dev` to be
safe. This is a workaround for not being able to manage package
versions specifically enough.