Remove the 'high disk time' alert.
This is a [cause-based-alert] and has not proved useful. We have disk time
graphs on dashboards.

In the unlikely event that this leads to a gap in our alerting coverage,
that gap should be filled by adding an alert on the user-facing response
time rather than alerting on an internal detail like disk I/O time.

[cause-based-alert]: https://docs.google.com/document/d/199PqyG3UsyXlwieHaqbGiWVa8eMWi8zzAn0YfcApr8Q/preview#heading=h.6ammb5h32uqq