Add Whitehall back into govuk_jenkins::jobs::integration_deploy::applications
Whitehall was removed from the Carrenza
govuk_jenkins::jobs::integration_deploy::applications as part of #10178
(during AWS migration).

However, this list is used to populate the integration-app-deploy
Jenkins job, which runs in CI (still in Carrenza).  Therefore we were
unable to deploy Whitehall to integration.  A comment at the top of this
list also states that migrated apps must be included here.

Adding this value back into the Carrenza hieradata.