Delay before alerting on unicorn workers
Previously we notified about abnormal unicorn workers immediately,
which is often a false alarm and due to an app restart. This changes
the check to delay by 2 minutes (2 * check_interval), in the case
the check runs near the time of the restart.

Making this change increases the minimum alert delay to 2 minutes,
and the maximum alert delay to 7 minutes (default_interval + 2).
This seems like a reasonable compromise.