Disable integration data for Whitehall content
This disables parts of the prod/staging -> integration sync, so that
the environment remains unchagned while the budget is being prepared.

While the DB sync for Whitehall is now done via govuk_env_sync, the
associated assets are still handled via env-sync-and-backup, so it's
still necesssary to disable the copy_data_to_integration job.