Use different reboot locks for different tasks
Previously, the with_reboot_lock script would attempt to create
/etc/unattended-reboot/no-reboot/reboot_lock when it didn't exist. I
think the intention was to avoid the machine from rebooting, but this
also has the effect of serialising anything using this script.

I'm been looking at the use of this for the govuk_env_sync cron
jobs. They don't need to run one after the other, but have been doing
due to this script.

Rather than using a single lock file, have the with_reboot_lock script
take the filename as the first argument. This commit also updates
where this script is used in govuk-puppet.