Add a jenkins task to publish the coronavirus page
Special-route-publisher only exists to be run as a jenkins task, so it
doesn't need deploying after changes are merged. This PR adds two tasks,
one to publish the page as a major update and one as a minor update.

It's likely these will be removed if the page is subsumed into a publishing
app.