Change Varnish threading configuration
Increase the minimum number of threads, with the intention of being
able to handle spikes in traffic more easily.

Increase the max threads value to again hopefully better handle large
spikes in traffic. Note that the documentation I'm reading says 5000
is probably the highest this value should be set to [1].

1: https://book.varnish-software.com/3.0/Tuning.html#number-of-threads