Rate limit specific high-impact Google Apps script
We are seeing spikes of load on the search-api as a specific
GApps script makes up to multiple thousands of requests over a couple of
minutes. The script is using multiple IPs and thus bypassing our
conventional load limiting.

However they all share the same app type in the user agent header, which
allows us to track which requests are being made from this one source.

Unfortunately this event is regular enough that we have decided to take
direct action.

This pull request attempts to create a zone that handles
Google-Apps-Script user agents specifically, to avoid such high spikes
from third parties.