Silence Icinga warnings in integration on secondary mongo machines.
Icinga has been showing warnings in integration that the various push
and pull jobs managed by `govuk_env_sync` had no run recently enough.

The reason behind this is that the script was exiting whenever a job
tried to run on a non-master database machine, this resulted in the
script never sending a nagios OK to Icinga, causing the freshness
warnings to eventually trigger.

This change removes the exit condition and adds some logging, so
when the jobs attempt to run on the non-master machine they just skip
over that machine, and that event is logged, that way a nagios OK is
still sent, and the logging makes it obvious why the jobs don't appear
to do anything on a secondary machine, as they cannot without write
access.