Specify aptly GPG key in hiera
The key was, as far as I can tell, manually added to the Carrenza
apt-1 machine years ago.  It's not there in AWS.  So, we need to start
managing this through puppet.

Step 1 is this change.  The secret key will also need adding to
govuk-secrets.  I've added Carrenza hieradata for this because the key
is still there in Carrenza, so it's easier than working out if that's
safe to delete.

Step 2 will be changing everywhere which has the fingerprint
hard-coded to use `hiera('apt_mirror_fingerprint')` instead.