Remove deprecated 'SES alert'
The HMGGDS SES account was once used for more things - but now
it's used for publishing apps' emails - and we're moving away
from it towards Notify. This alert has not fired in Production
AWS for 342 days, and in Carrenza for 83 days.

The checks report that we're using 0.154% of our daily SES quota,
which really is not going to be a problem.