Set asset paths to allow more connections
This make the connection and burst limits consistent for asset paths as
the settings on the asset host name. I've done the somewhat lazy thing
here of copying the values over without calculating the exact maths as
to whether these numbers reflect demand.

I've gathered from the comments on the previous commits that these
numbers don't have a whole lot of impact because most of our asset
requests are cached. However, I want to create as little divergence as
possible in this switch over so prefer to keep them as consistent.