Allow multi asset pipeline prefixes
Changing from one asset path to another means that we have a period of
time where we serve assets with incorrect cache headers and need to
carefully co-ordinate deployments. As part of RFC 115 [1] we are going
to be transitioning these app paths so they can be hosted on the
www.gov.uk hostname.

To provision for this swap over I've switched the pipeline prefix argument
to be an array so that it can accept multiple items, this will allow both
configurations to be in use as at a particular time.

The downside of this is that the argument is a bit confusing when it
will nearly always be used just for a single item.

[1]: https://github.com/alphagov/govuk-rfcs/blob/45e09c088bfde6c1d4dfb68f9bc1c635f63c9a39/rfc-115-enabling-http2-on-govuk.md