Use live URL when draft is missing
`draft_content_store` sets `PLEK_HOSTNAME_PREFIX` to prepend `draft-` to the
URIs of its dependencies, so that it uses the special "draft" versions
of those dependencies. Two of these dependencies draft-info-frontend and
draft-feedback, have been causing errors since their draft version is
is missing.

To prevent the errors we overridden the URIs to their live
versions in draft_content_store.

trello card: https://trello.com/c/LNbGiOcM/1898-5-investigate-an-interesting-sentry-error-relating-to-the-router