Make 'integration-app-deploy' job generic
Previously we had an 'integration-app-deploy' job that was triggered by
a successful build of the master branch for each app on CI. This renames
the job to 'Deploy_App_Downstream' and extracts it so that it can be run
in other environments, as part of our move to Continuous Deployment.

In order for this job to run safely in other environments, this adds
prepends a new step to ensure that the commit being deployed really is
the latest release of the master branch for the repo in question. Due to
historical issues with the naming of some repos, we need to do a lookup.

Note that, although the job is configured with all apps in the Carrenza
environment, it will only be provisioned in the CI environment for that
provider. For all other environments, it will only be provisioned in
AWS, since most apps are now deployed in AWS.