Enable continuous deployment to Staging
Previously we had an 'integration-app-deploy' job that was triggered by
a successful build of the master branch for each app on CI. This renames
the job to 'Deploy_App_Downstream' and modifies the 'Deploy_App' job to
call it after a successful run of Smokey i.e. Continuous Deployment.

Continuous Deployment (CD) was originally proposed in RFC 70 [1]. Manual
deployments to staging and then production take up a lot of time. Although
we have Smokey tests, we still rely on manual testing for many deployments,
which means 'full' CD would be quite a challenge. Enabling CD to Staging
is a step in that direction, which will have a few benefits:

- It reduces the amount of time spent deploying things
- It encourages us to improve our Smokey tests
- It encourages us to not ignore our Smokey tests

While there is a risk of 'broken' code being deployed to Staging, this
is also a risk now: sometimes we don't see the warnings [2]. In general,
we shouldn't be deploying 'broken' code; and if that does happen, we
should have automated (Smokey) tests to protect us from ourselves,
instead of relying on (people remembering to do) manual testing.

[1]: https://github.com/alphagov/govuk-rfcs/blob/master/rfc-070-path-towards-continuous-deployment-cd.md
[2]: https://github.com/alphagov/smokey/pull/633